Changed the database and fixed switching between screens
As you probably realised by now our storage exceeded maximal file transfer rate for a day of 1 GB/day. As there are only 2 days to the presentation, I made a new firebase project and changed the firebase.js in the process. I have also added several precautions to make the app not to make as many fetch requests as before. The app now initialises with welcome page instead of explore, therefore it will not fetch 20 or some images as it starts. Even so please upload as little items as you possibly can during development. Thank you